🚚 [Rename]: PlaceListResponse -> PlacePreviewResponse
✨ [Feat]: DTO 생성을 Converter로 위임에 따라 필요 속성들 수정
